Restructure devices into single files
