+++ functions/devices/awning.js
@@ -1,0 +1,9 @@
+const OpenCloseDevice = require('./openclosedevice.js');
+
+class Awning extends OpenCloseDevice {
+  static get type() {
+    return 'action.devices.types.AWNING';
+  }
+}
+
+module.exports = Awning;
