+++ functions/devices/door.js
@@ -1,0 +1,9 @@
+const OpenCloseDevice = require('./openclosedevice.js');
+
+class Door extends OpenCloseDevice {
+  static get type() {
+    return 'action.devices.types.DOOR';
+  }
+}
+
+module.exports = Door;
