+++ functions/devices/garage.js
@@ -1,0 +1,9 @@
+const OpenCloseDevice = require('./openclosedevice.js');
+
+class Garage extends OpenCloseDevice {
+  static get type() {
+    return 'action.devices.types.GARAGE';
+  }
+}
+
+module.exports = Garage;
