+++ functions/devices/gate.js
@@ -1,0 +1,9 @@
+const OpenCloseDevice = require('./openclosedevice.js');
+
+class Gate extends OpenCloseDevice {
+  static get type() {
+    return 'action.devices.types.GATE';
+  }
+}
+
+module.exports = Gate;
