+++ functions/devices/hood.js
@@ -1,0 +1,9 @@
+const Fan = require('./fan.js');
+
+class Hood extends Fan {
+  static get type() {
+    return 'action.devices.types.HOOD';
+  }
+}
+
+module.exports = Hood;
