+++ functions/devices/outlet.js
@@ -1,0 +1,9 @@
+const Switch = require('./switch.js');
+
+class Outlet extends Switch {
+  static get type() {
+    return 'action.devices.types.OUTLET';
+  }
+}
+
+module.exports = Outlet;
