+++ functions/devices/simplefan.js
@@ -1,0 +1,9 @@
+const Switch = require('./switch.js');
+
+class SimpleFan extends Switch {
+  static get type() {
+    return 'action.devices.types.FAN';
+  }
+}
+
+module.exports = SimpleFan;
