+++ functions/devices/simplehood.js
@@ -1,0 +1,9 @@
+const Switch = require('./switch.js');
+
+class SimpleHood extends Switch {
+  static get type() {
+    return 'action.devices.types.HOOD';
+  }
+}
+
+module.exports = SimpleHood;
