+++ functions/devices/startstopswitch.js
@@ -1,0 +1,26 @@
+const DefaultDevice = require('./default.js');
+
+class StartStopSwitch extends DefaultDevice {
+  static get traits() {
+    return [
+      'action.devices.traits.StartStop'
+    ];
+  }
+
+  static get requiredItemTypes() {
+    return ['Switch'];
+  }
+
+  static getState(item) {
+    let state = item.state === 'ON';
+    if (this.getConfig(item).inverted === true) {
+      state = !state;
+    }
+    return {
+      isRunning: state,
+      isPaused: !state
+    };
+  }
+}
+
+module.exports = StartStopSwitch;
