+++ functions/openhab.js
@@ -115,4 +115,4 @@ class OpenHAB {
 	}
 }
 
-module.exports = { OpenHAB };
+module.exports = OpenHAB;
