+++ functions/package.json
@@ -12,6 +12,7 @@
   ],
   "license": "EPL-2.0",
   "dependencies": {
-    "actions-on-google": "^2.12.0"
+    "actions-on-google": "^2.12.0",
+    "glob": "^7.1.6"
   }
 }
