+++ package.json
@@ -26,6 +26,11 @@
   },
   "devDependencies": {
     "@types/jest": "^24.9.1",
+    "eslint-config-standard": "^14.1.1",
+    "eslint-plugin-import": "^2.21.2",
+    "eslint-plugin-node": "^11.1.0",
+    "eslint-plugin-promise": "^4.2.1",
+    "eslint-plugin-standard": "^4.0.1",
     "jest": "^24.9.0"
   }
 }
