+++ tests/devices.metadata.test.js
@@ -1,4 +1,5 @@
 const Devices = require('../functions/devices.js');
+const Thermostat = require('../functions/devices/thermostat.js');
 
 describe('Test Switch Devices with Metadata', () => {
   test('Switch Type', () => {
@@ -416,7 +417,7 @@ describe('Test Thermostat Device with Metadata', () => {
   });
 
   test('usesFahrenheit', () => {
-    expect(Devices.Thermostat.usesFahrenheit({
+    expect(Thermostat.usesFahrenheit({
       metadata: {
         ga: {
           value: 'Thermostat',
@@ -429,7 +430,7 @@ describe('Test Thermostat Device with Metadata', () => {
   });
 
   test('getAttributes', () => {
-    expect(Devices.Thermostat.getAttributes({
+    expect(Thermostat.getAttributes({
       metadata: {
         ga: {
           value: 'Thermostat',
@@ -443,7 +444,7 @@ describe('Test Thermostat Device with Metadata', () => {
       thermostatTemperatureUnit: 'F',
     });
 
-    expect(Devices.Thermostat.getAttributes({
+    expect(Thermostat.getAttributes({
       metadata: {
         ga: {
           value: 'Thermostat',
@@ -457,7 +458,7 @@ describe('Test Thermostat Device with Metadata', () => {
       thermostatTemperatureUnit: 'C',
     });
 
-    expect(Devices.Thermostat.getAttributes({
+    expect(Thermostat.getAttributes({
       metadata: {
         ga: {
           value: 'Thermostat',
@@ -472,7 +473,7 @@ describe('Test Thermostat Device with Metadata', () => {
     });
 
     // test thermostatTemperatureRange attribute
-    expect(Devices.Thermostat.getAttributes({
+    expect(Thermostat.getAttributes({
       metadata: {
         ga: {
           value: 'Thermostat',
@@ -491,7 +492,7 @@ describe('Test Thermostat Device with Metadata', () => {
     });
 
     // wrong value for thermostatTemperatureRange should not throw error
-    expect(Devices.Thermostat.getAttributes({
+    expect(Thermostat.getAttributes({
       metadata: {
         ga: {
           value: 'Thermostat',
@@ -506,7 +507,7 @@ describe('Test Thermostat Device with Metadata', () => {
     });
 
     // wrong value for thermostatTemperatureRange should not throw error
-    expect(Devices.Thermostat.getAttributes({
+    expect(Thermostat.getAttributes({
       metadata: {
         ga: {
           value: 'Thermostat',
@@ -522,7 +523,7 @@ describe('Test Thermostat Device with Metadata', () => {
   });
 
   test('getModeMap', () => {
-    expect(Devices.Thermostat.getModeMap({
+    expect(Thermostat.getModeMap({
       metadata: {
         ga: {
           value: 'Thermostat'
@@ -538,7 +539,7 @@ describe('Test Thermostat Device with Metadata', () => {
       'eco': ['eco']
     });
 
-    expect(Devices.Thermostat.getModeMap({
+    expect(Thermostat.getModeMap({
       metadata: {
         ga: {
           value: 'Thermostat',
@@ -554,7 +555,7 @@ describe('Test Thermostat Device with Metadata', () => {
       'cool': ['cool']
     });
 
-    expect(Devices.Thermostat.getModeMap({
+    expect(Thermostat.getModeMap({
       metadata: {
         ga: {
           value: 'Thermostat',
@@ -573,7 +574,7 @@ describe('Test Thermostat Device with Metadata', () => {
   });
 
   test('translateModeToGoogle', () => {
-    expect(Devices.Thermostat.translateModeToGoogle({
+    expect(Thermostat.translateModeToGoogle({
       metadata: {
         ga: {
           value: 'Thermostat',
@@ -584,7 +585,7 @@ describe('Test Thermostat Device with Metadata', () => {
       }
     }, 'COMFORT')).toBe('heat');
 
-    expect(Devices.Thermostat.translateModeToGoogle({
+    expect(Thermostat.translateModeToGoogle({
       metadata: {
         ga: {
           value: 'Thermostat',
@@ -597,7 +598,7 @@ describe('Test Thermostat Device with Metadata', () => {
   });
 
   test('translateModeToOpenhab', () => {
-    expect(Devices.Thermostat.translateModeToOpenhab({
+    expect(Thermostat.translateModeToOpenhab({
       metadata: {
         ga: {
           value: 'Thermostat',
@@ -608,7 +609,7 @@ describe('Test Thermostat Device with Metadata', () => {
       }
     }, 'heat')).toBe('COMFORT');
 
-    expect(Devices.Thermostat.translateModeToOpenhab({
+    expect(Thermostat.translateModeToOpenhab({
       metadata: {
         ga: {
           value: 'Thermostat',
@@ -621,7 +622,7 @@ describe('Test Thermostat Device with Metadata', () => {
   });
 
   test('getMetadata', () => {
-    expect(Devices.Thermostat.getMetadata(
+    expect(Thermostat.getMetadata(
       {
         name: 'MyItem',
         label: 'MyThermostat',
@@ -703,7 +704,7 @@ describe('Test Thermostat Device with Metadata', () => {
   });
 
   test('getState', () => {
-    expect(Devices.Thermostat.getState({
+    expect(Thermostat.getState({
       type: 'Group',
       metadata: {
         ga: {
@@ -712,7 +713,7 @@ describe('Test Thermostat Device with Metadata', () => {
       }
     })).toStrictEqual({});
 
-    expect(Devices.Thermostat.getState({
+    expect(Thermostat.getState({
       type: 'Group',
       metadata: {
         ga: {
@@ -753,7 +754,7 @@ describe('Test Thermostat Device with Metadata', () => {
       'thermostatMode': 'off'
     });
 
-    expect(Devices.Thermostat.getState({
+    expect(Thermostat.getState({
       type: 'Group',
       metadata: {
         ga: {
