+++ tests/openhab.test.js
@@ -1,5 +1,4 @@
-
-const OpenHAB = require('../functions/openhab.js').OpenHAB;
+const OpenHAB = require('../functions/openhab.js');
 
 describe('Test EXECUTE', () => {
   test('OnOff Switch', async () => {
